Check if two vectors contain anagrams
Difficulty: Easy

Given two vectors, compare each containing word of <vector1> to the one
on the same position in <vector2>.

I used additions to obtain the result. This algorithm is case-sensitive
(it is quite easy to change, but I'm lazy)

Note: This is NOT the most optimal solution, it's just an easy, fast
approach that can be easily improved.